requirements.txt
Flask==2.0.2
PyPDF2==1.26.0
pdf2image==1.16.0
pdfminer==20191125
Pillow==8.4.0
PyPDF2==1.26.0
numpy==1.19.5
pngquant==1.0.7
pytesseract==0.3.8
scipy==1.5.2
scikit-learn==0.24.2
ocrmypdf==12.7.2
tqdm==4.64.1
Werkzeug==2.0.2
xkcdpass==1.19.3
